refactored code into util classes, for future reusability
